phase-II/backend/pyproject.toml
[project]
name = "fastsapi-crud"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.124.4",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.27",
]
